Refactor a few functions and move them util. Add new parameter sequencing file.
