+++ rotary/sample.py
@@ -35,8 +35,10 @@ class SequencingFile(object):
         r_value = short_read_r_regex.findall(remaining)
         if r_value:
             self.r_value = r_value[0].upper()
+            self.short = True
         else:
             self.r_value = None
+            self.short = False
 
 
 class Sample(object):
@@ -183,20 +185,6 @@ def create_sample_tsv(output_dir_path, samples):
             tsv_writer.writerow(current_sample.sample_file_row)
 
     return sample_tsv_path
-
-def file_is_gzipped(file_path):
-    """
-    Determine if a file is gzipped based in the file extension.
-
-    :param file_path: The path to the file to be checked.
-    :return: True if the file is gzipped, False otherwise.
-    """
-    extension = os.path.splitext(file_path)[1]
-
-    if extension == '.gz':
-        return True
-    else:
-        return False
 
 
 def find_samples_in_fastq_directory(input_path):
