+++ rotary/utils.py
@@ -8,8 +8,6 @@ Description: Utilities for the Rotary hybrid assembly workflow.
 import gzip
 import os
 
-from rotary.sample import file_is_gzipped
-
 
 def check_for_files(output_dir_path, files_to_check):
     """
@@ -100,3 +98,18 @@ def symlink_or_compress(in_file_path, out_file_path):
         os.symlink(in_file_path, out_file_path)
     else:
         gzip_file(in_file_path, out_file_path)
+
+
+def file_is_gzipped(file_path):
+    """
+    Determine if a file is gzipped based in the file extension.
+
+    :param file_path: The path to the file to be checked.
+    :return: True if the file is gzipped, False otherwise.
+    """
+    extension = os.path.splitext(file_path)[1]
+
+    if extension == '.gz':
+        return True
+    else:
+        return False
